ops_backend/requirements.txt
# Django Requirements
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7

# PDF Generation
WeasyPrint==60.2
cffi==1.16.0

# HTTP Client
requests==2.31.0

# Security
pip-audit==2.6.1

# Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Redis
redis==5.0.1
django-redis==5.4.0
